Fixes dropbox download to handle non-zip/non-rar file downloads
Currently there is an error raised if downloading a non-zip or non-rar file. It will still download the file, but it will append a '.zip' after the extension name. By adding a check to see if the file is indeed a rar/zip file, we can only run unzip if it is. Otherwise leave it alone and the file downloads correctly :)